- En el operador de indexación de GroupLines el id del elemento se pasa como size_t - Console: Propiedades de fuente

+++ src/geometric_entities/segment.h
@@ -599,8 +599,8 @@ public:
    * \param[in] id Indice del elemento
    * \return Linea
    */
-  const Line &operator[](int id) const { return linesgroup[id]; }
-  Line &operator[](int id)  { return linesgroup[id]; }
+  const Line &operator[](size_t id) const { return linesgroup[id]; }
+  Line &operator[](size_t id)  { return linesgroup[id]; }
 
   const std::vector<Line> &getLines() const { return linesgroup; }
